% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tour.R
\name{run_tour}
\alias{run_tour}
\title{Run tour}
\usage{
run_tour(d, m, labels = NULL)
}
\arguments{
\item{d}{a nx2 data frame with columns 'x' and 'y' for the static plot}

\item{m}{a nxp numeric matrix for touring in 2 dimensions}

\item{labels}{optional a character/factor vector of labels (classes) for each obervation.}
}
\description{
Run tour
}
\examples{
\dontrun{

library(mvtnorm)
s <- matrix(c(1, .5, .5, .5, 1, .5, .5, .5, 1), ncol = 3)
m <- rmvnorm(500, sigma = s)
d <- dmvnorm(m)
# take 1D slice of the 3D density
dat <- data.frame(x = m[, 1], y = d)
run_tour(dat, m)
}


}

