% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourbrush.R
\name{tourbrush}
\alias{tourbrush}
\title{tourbrush}
\usage{
tourbrush(df, aps = 2, palette = "Dark2")
}
\arguments{
\item{df}{a data frame}

\item{aps}{angles per second}

\item{palette}{an RColorBrewer palette}
}
\description{
tourbrush
}
\examples{
\dontrun{
library(mvtnorm)
s <- matrix(c(1, .5, .5, .5, 1, .5, .5, .5, 1), ncol = 3)
m <- rmvnorm(500, sigma = s)
# take 1D slice of the 3D density
d <- setNames(data.frame(m), c("x", "y", "z"))
d$density <- dmvnorm(m)
tourbrush(d)

vars <- c("pitch_type", "start_speed", "spin_dir", "break_y",
  "break_angle", "break_length")
tourbrush(yu[vars])


}
}

