% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yu}
\alias{yu}
\title{Pitches thrown by Yu Darvish}
\format{A data frame with many variables. See this post for a description of them
\url{https://fastballs.wordpress.com/2007/08/02/glossary-of-the-gameday-pitch-fields/}}
\usage{
yu
}
\description{
Pitch f/x data of every pitch thrown by Yu Darvish in 2013
}
\keyword{datasets}

